Functions `math.ceil()` and `math.floor()` return `int` in Python [for https://codeforces.com/contest/1521/submission/117139834]

+++ tests/python_to_cpp/Shed Skin Examples/2.sudoku2.py
@@ -101,7 +101,7 @@ class board:
                 self.cels[cr][cc] = self.cels[cr][cc] & mask
 
         def cell(self, num): # const
-                return int(math.ceil((num + 1) / 3.0)) - 1
+                return math.ceil((num + 1) / 3.0) - 1
 
         def to_str(self): # const
                 ret = ""
